Implement component svc, styles, view composition
Do some more renaming and implement a component registry that picks up
the rest of our component resources.

The scripts were already included in the js bundle created by gulp. We
now use the resource root infrastructure (to which we add a servlet
context resource based implementation) to integrate the dynanmic CSS
engine DSS implemented alongside OpenSAGA.

App.json
--------

We introduce an app.json file which contains the application-level
configuration values for the application.

The corresponding properties are added to the main application model
now just called "Application".

The ModelCompositionService keeps merging all the models from other
locations into the application model, only now using ignored props.

Styling
-------

We determine the styles used for every application from two sources:

We load all .css files whose names correspond to css definitions within
the same component package.

We also load stylesheets from an "styleSheets" string array in app.json.

These must include all base styles including bootstrap.css which it
contains by default in the base models.

View-Data / -Rendering
----------------------

We invoke the view data service to build all component data and provide
that data and the current view model to the client-side.

Turns out we need two different JSON representations:

We have the somewhat lax format we accept when reading and I initially
thought we would only write it out the same way again, stripping
generated ids etc.

But for the client side we need those ids. For now we add an option
to the Attributes jsonifier to control if we want generated ids or not.
(will revisit that when writing out models to disk)

We implement a view-service that renders the JSON view data into
readable scripts that render that view and merge the static view props
with the data received from the view data service.

We compile that source into react components and cache references within
the view service.

We will always render out id props for all proper react components and
it is up to the component to decide if that id is mirrored as HTML id.

For the HTML tags we render, we only output the ids if the id is not
generated.

Service layer configuration
---------------------------

We move the component scan of the service package to its own
ServiceConfiguration class.