Refactor resource handling, add hot-reloading/push
We implement general hot-reloading based on the java.nio watchers and
add a new controller that does long-poll waiting for change
notifications.

These change notifications consist of the client receiving either a
change model JSON or a special change model.

The special change models are:

 * de.quinscape.exceed.model.change.Timeout

   Signals that we reached the maximum time we want to let an long poll
   unanswered and causes the client just to start another long poll.

 * de.quinscape.exceed.model.change.StyleChange

   One of the stylesheets for the current application has changed and
   the client will refresh the linked style by adding a matrix-var
   timestamp to the end of the URL which the servlet container ignores
   for us.

 * de.quinscape.exceed.model.change.CodeChange

   The library source location was defined by setting the system
   property exceed.library.source and the base code was loaded via file
   system and was now updated.

   You can of course just muck around in the generated bundle, but the
   very clear recommendation is to have "gulp watch" run in the source
   directory.

   Gulp watch recompiles all changed source to the new bundle and then
   we reload the bundle by having the client re-request the code URL

 * de.quinscape.exceed.model.change.Shutdown

   Is sent by the application service when it the spring context is
   shut down.

   Current client-side implementation is to give up long-polling and
   notify the user of the stopped server with an alert asking them to
   reload the page.

   We might get more sophisticated later.

Resources
---------

We restructure our extension resources which so far have only included
models to also include other application resources. Basically everything
but the base JSP views which just work since the servlet API handles
the resource overriding for us.

The rest of the resources, including the compiled js code is now handled
by a resource controller mapped to /res/{appName}/** where ** is the
resource path within our restructured extensions. We remove the
ScriptController and ScriptService.

We update to react 0.14.0 which causes us to have a second shimed
library ("react-dom" -> ReactDOM global).