Component Attributes: Parse and cache expressions
With all the hotloading going on it's easy to get side-tracked by all
the dynamic stuff going on and forget the basics.

I think it's important to tune the design of the application towards a
good final runtime performance in a future production mode execution.

As far as possible we need to take advantage from the model and regard
it as conceptually constant and static.

We then have to make the dev run eat the performance penalties involved
with editing the application, which means often updating larger portions
then needed for the update to work with the intention of keeping the
code structure simple and clean.

We always load a model file as complete model and remerge that with the
application model. We reload the component registration completely when
resources change.

So for the attributes that might have expressions associated with them
we cache the resulting expression AST in the model since it's basically
just another kind of static model.

This means we only ever have to reparse the expressions if we reload
the whole model.