Implement resource caching
Make the resource loader cooperate with an injected guava loading cache.
If such a cache is set, register a resource change listener that routes
changes to the .refresh() method of the cache.

We create an new class to be cached named CachedResource which contains
the byte[] data and an UUID which we use as ETag source for the resource
controller

Add a ResourceCacheFactory interface to create the cache (
LoadingCache<String,CachedResource> ) and a DefaultImplementation that
does this based on weighed the byte[] data in KB and a configurable
per-application cache limit.

The caches are per application which is not really ideal because it will
cache the base resources over and over again. But then again those
resources shouldn't be too big in production and we might just want to
run with different base versions in the future.