New Concept: Expression Environment
So far we have been dealing with expressions from a purely syntactic
point of view. We chose a single unified expression language for ease of
implementation and interchangeability of AST nodes and tokens etc.

The usage of the expressions will vary a lot overall and so do the
identifiers and functions present.

In most cases we will just interpret the expression or transform it into
another format.

Model validation rules describe inter-model relations and also act as
completion source.

Query expressions decribe the (potentially joined) querying of domain
objects etc.

Expression Environment
----------------------

We introduce a new abstract base class for the different use cases of
our expressions. Each environment is an AST visitor with a default
implementation for common operations and protected flags to allow /
disallow their presence in expressions of a certain kind.

Identifiers have different types and meanings in the expressions, if
present at all.

For example a query expression like this

 Foo.join(Bar).on(Foo.bar_id == Bar.id)

needs to generally accept identifiers and provide values for "Foo" and
"Bar"

.join() and .on() are operations applied onto the values in the
expression environment.

To ease implementation, we use a reflective approach to resolving
operations within an expression environment.

Here is a test environment as example:

public class TestEnvironment extends ExpressionEnvironment
{
    public TestEnvironment()
    {
        comparatorsAllowed = true;
        arithmeticOperatorsAllowed = true;
        logicalOperatorsAllowed = true;
    }

    @Override
    protected Object resolveIdentifier(String name)
    {
        switch (name)
        {
            case "num":
                return 123;
            case "value":
                return "abc";
            case "value2":
                return "xyz";
            case "TRUE":
                return true;
            case "FALSE":
                return false;
            default:
                throw new ExceedRuntimeException("Unknown identifier: "+ name);
        }
    }

    public Foo foo(ASTFunction node)
    {
        return new Foo();
    }

    public Foo name(ASTFunction node, Foo foo)
    {
        foo.setName((String) visitOneChildOf(node, ASTString.class));
        return foo;
    }
}

The constructor allows comparators, logical and arithmetic operators
which would just throw an exception if disallowed

the overloaded resolveIdentifier method provides values for the
available idenitifiers and throws if unavailable.

We also see two operations being defined foo() and name().

An operation method must have ASTFunction as first parameter and have a
potential second argument which is the context to which it can be
applied.

foo() produces an Foo POJO instance value and name() then is an
operation that can be applied to this Foo typed context, in this case
setting the name property.

	foo().name('Bla')

would be an expression using both.

The supplied function node contains the AST nodes for the operation
arguments.

We use above test environment to test general operation and add some
error cases to the expression parser test making sure we only deal
with reasonable number of operands.