Query Expression Environment
The query expression environment is the first major implementation of
the expression environment abstraction.

Context: Component Queries
--------------------------

So we have components defined as a component package directory
containing Javascript sources, stylesheets and also the component JSON
descriptor that describes the component available in the package.

One of the definitions are component queries. In the earlier prototypes
these were just JSON blocks, but it seems much preferable to have query
expressions instead.

{
    "components": {
        "DataGrid": {

            ...

            "queries": {
                "result": "<QUERY_EXPRESSION>"
            },

            ...
}

Each component can define a "queries" block that contains a map which
maps query names to the query expressions for them.

Each of the query names will become a React component prop unless that
component defines a static prop overriding that (in which case the query
is never executed).

The QueryDataProvider then uses the expressions to drive a JOOQ SQL
query.

The plan is to use the query expressions as primary means of data
acquisition.

In addition to that there are two ways to change that:

 * Implement an alternate dataprovider
 * Implement a different QueryExecutor to execute the semantically same
   query based of another data source.

The queries are able to provide joined, filtered and paged domain object
results.

In spirit with the general dynamism of everything and operation via
generated JOOQ classes being optional, we embrace the data output of
the query expression model as general data format.

Query Result Format
-------------------

For every query expression, the component receives a query result block
which looks like this

{
    "entities": [
        {
            "type": "Foo",
            "pkFields": ["id"]
        }
    ],
    "fields": {
        "id": {
            "type": "UUID",
            "maxLength": 36,
            "entityIndex": 0,
            "name": "id",
            "qualifiedName": "AppUser.id"
        },
        "name": {
            "type": "PlainText",
            "maxLength": 64,
            "entityIndex": 0,
            "name": "name",
            "qualifiedName": "AppUser.name"
        }
    },
    "rows": [
        {
            "id": "150a9ba8-e612-421c-a458-e43e3e608ac7",
            "name": "Foo #1"
        }, {
            "id": "35b60665-388a-4718-a1f6-238cb1c700b3",
            "name": "Foo #2"
        }, {
            "id": "b7d8275c-d759-460f-8119-ef7d3032d1a7",
            "name": "Irregular Foo"
        }
    ]
}

The "entities" property lists all domain entities involved in the query
and their primary key fields.

The "fields" property maps the local field name to a domain property
like structure that has all the relevant fields of domain property and
also has two more properties:

 * entityIndex points to the entity the field belongs to
 * qualifiedName is the full qualified field name within the application

The "rows" property than contains the data payload. It is a list of map
objects mapping a local field name to the value in the current
row (converted using the corresponding "type" from the "fields" map).

Query Expression Syntax
-----------------------

The query expression syntax in general mirrors the features provided by
SQL / JOOQ.

The defined domain types in the application are available as table
identifiers. These can be joined together.

	Foo.as(f).join(Bar).on(f.bar_id == Bar.id)

Here a simple join over a foreign key field.

The as operator can be used to rename / disambiguate fields. All fields
can be defined unqualified if that reference is unambiguous. They can
also be given as <type>.<property-name> or as <alias>.<property-name>

The potentially joined tables can then be queried.

	query(Foo).filter(Foo.type == 2).limit(10).orderBy(Foo.name)

The query operation produces a query definition object which then can be
configured with

 * filter: filters the list
 * limit: limits the number of results
 * offset: number of rows to skip
 * orderBy: list of field names

The order of these operations does not matter as they overall describe a
query before it is being executed, not a live-stream of objects.

The expressions can use var() and prop() functions to refer to component
props or vars.

Here for example the definition for the datagrid:

    query(
        prop('type')
            .fields( childModelNames())
        ).filter(prop('filter'))
        .orderBy(var('orderBy'))
        .limit(var('limit'))
        .offset(var('offset'))

The query is completely composed out of values from the current runtime
context or model

prop('type') imports the "type" component attribute as expression into
the expression, similar for filter.

.orderBy() used the current value of the "orderBy" var etc.

qr fixup