DomainObject: Refactor hierarchy
While we want in the end support generating jOOQ POJOs out of our domain
type definitions, the system has to be able to run totally dynamically
on JSON definitions.

We turn DomainObject into an interface and add generic set property, get
property and list propery methods move the default methods into. We
create a new class AbstractDomainObject which then is extended by two
implementations:

 * GenericDomainObject

   Is basically just a typed property map based on svensons
   DynamicProperties.

 * GeneratedDomainObject

   Is an intermediary base class for future generated POJOs. We use
   svenson JSON property util to dynamically access properties over the
   property methods that the domain object interface copies from
   DynamicProperties.

We will be using the GenericDomainObject a lot as container for joined
field rows. We will need to make that work for pojos, too. Basically a
set of POJOs flattened into one property namespace.