Fix Hotloading, caching , config, components
This is a bit of a leftovers commit I'm no longer willing to spend time
on separating from one another. It ties up the loose end of the tickets
before this and takes care of making hot reload work again, use the
resource caching and resource controller correctly and changes the
configuration classes to reflect the current changes.

It contains a first simple design for a toggable sidebar in-page editor
compoment and the first version of the datagrid doing simple domain
queries. Needs more fixing to return to previous prototype
functionality.

Component Model descriptors
---------------------------

Prepare general ComponentDescriptor and PropDeclararation containers,
change to expression based format.

Query expressions for datagrid already ready to work, others just
sketches of future things to come.

Component Model Context
=======================

The views are composed at runtime from the component models by
ViewComponent.js.

What we were still missing a mechanism to aid composability. We want to
support something like this JSX example

	<DataGrid ...>
		<DataGrid.Column name="foo">
			<b>{ context }</b>
		</DataGrid.Colum>
	</DataGrid>

The DataGrid defines its query and the DataGrid.Column refers to one of
to a field "foo" in that query. The context variable now is supposed to
contain the value of the field "foo" in the current row.

This could be done with react contexts I think, but this.context is not
accessible from the outer instance composing the datagrid.

We help us by introducing a new concept for a component.

Context-Provider
----------------

If a component descriptor contains a "contextProvider" == true property,
we just do the normal rendering until we reach that component. The
component receives an "renderChildrenWithContext" function prop.

The component can render whatever it wants and then call
this.props.renderWithContext with the context object. This context
object will be made available to the children in two ways:

 * this.props.context
   As context prop of each child
 * As "context" identifier in a component model attribute expression

Context-Key
-----------

As a second mechanism, a component can define a component descriptor
property "contextKey" containing the name of an attribute.

When the component receives its context, it will not receive the initial
context provided, but a property of this context as defined by the
attribute value.

We use this for DataGrid.Column to automatically go from the row context
provided by the datagrid to the appropriate property corresponding to
the DataGrid.Column name attribute.