Refactor Component registration and data fetching
We implement a component runtime registration / component data provider
/ query executor system.

We have our view model that largely consists of a tree of component
model instances with an JSON representation

Component Model
---------------

{
	"name" : "Bla",
	"attrs:" {
		"id": "myId",
		"value" : "{ foo }"
	}
	"kids" : [ ... ]
}

Each component model corresponds to a React component or React HTML tag.
We have already implemented a two-pronged system for dealing with the
component definitions.

We find all "component.json" descriptors in the component package
directory and require a CommonJS module for each Component declaration
we encounter (unless it's namespaced like "DataGrid.Column" which we
would expect to be also defined in the module DataGrid.js)

The json and js files are bulk-required into the Javascript bundle built
by the gulp build.

We also load all the corresponding styles into our styleservice (".css"
instead of ".js")

The name of our component model corresponds to the component names
declared in the component descriptors. We follow the react convention of
using uppercase names for user components and lowercase names for normal
HTML elements.

The attrs largely correspond to React component props. The values can
either be JSON literals or an expression which is a string wrapped in
braces. These attribute expressions are evaluated by the ViewComponent
when building the view out of component models.

The kids array contains the child component models. Each component has a
static component descriptor which is the JSON loaded into the
ComponentDescriptor class. There is also a runtime registration class
ComponentRegistration that is being kept in a component registry.

We keep a reference to the runtime component registration object for
every component model in the tree.

The runtime component registration links to the component name, the
static descriptor and data provider instance.

Data-Provider
-------------

The data provider system is one of two ways of customizing data fetching
for exceed components.

It's a simple interface that has one method that gets called for every
component instance with the data provider context and the runtime
variables for that component and which returns a data map for that
component. It can also use the data-provider context to only selectively
provide data for component sub-trees (think tab-set).

The component author can define an alternate data provide bean name in
the component descriptor as "dataProvider" property.

If no such property exists, the default data provider is used.

One prerequisite for using a data provider is having a component model
id attribute defined on the model. We had an auto-id generator in
OpenSAGA and I started implementing one here, too, but it seemed a lot
more trouble than it really helped and so far we don't really need model
ids in every case, just to leave data for a component under its id.

Query Data Provider
-------------------

The query data provider is the default data provider. It provides data
by executing query expressions defined under "queries". It uses the
QueryTransformer to transform the query expression string into a
QueryDefinition object graph.

This graph gets then handed to a query executor which executes the
actual data-fetching as defined by the query definition.

The default query executor is the JOOQQueryExecutor which executes query
definitions by transforming them into jOOQ DSL calls and conditions.

The "queryExecutors" property in the component descriptor can be used to
customize the query executors for a component per query.

It contains a map with query executor implementation bean names to use
instead of the default query executor.