Domain: Add Layout Model
We will need to store graphical meta-data for some of our models and
I've decided to keep it separate.

We use the same infrastructure as for models, the new Layout class is
also a TopLevelModel etc.

We use a resource location /layout/domain.json for the domain layout
data to really separate layout information from the model information as
such and to allow easy selective deleting.

We extend ModelCompositionService to merge in the domain layout class.