Editor Controller
We create an independent controller to serve JSON data to the runtime
editor on the client and to receive and store JSON from the client in
return.

We add one end point /editor/{app}/domain that serves all domainTypes,
all propertyTypes and the current domain layout as single JSON block.

We also create a POST endpoint /editor/{app}/domain/layout that can
store the layout for a single domain entity.

So far we just synchronize on the application model (which I think is a
good idea as standard behavior to prevent odd thread-safety issues.

We should write out the domain config in regular intervals.