Add SVGLayout helper
The support of the SVG API for text support is kind of limited. We don't
really know how big font-sizes are.

To have a rough estimate, we create a svg layout helper that does some
text measurements.

It creates an off-screen SVG element that contains a test phrase
containing all letters, mostly uppercase, but with lower case characters
if they have ascenders or descenders.

This way we determine SVG coordinate heights for three font-sizes which
we handle as symbolic enum LARGE,NORMAL,SMALL.

We also calculate an "pixel per character" value based on the SVG
coordinate with divided by the numbers of characters in the test phrase.

We use react for simplicity sake and rather accept the process being
asynchronous.

We make sure to initialze the layout module right at the start and wait
for promise resolution before doing the rest of the initialization.