Expression: Add map literals
Add map literals to the expression language to be able to express maps
as values and also as configuration tool for action model expressions.

We follow js and allow idenitifiers without quotes as keys (no computed
keys).

We enable all operations in var expressions because there's no good
reason to disallow them.