Actions: Unified Actions, server-side
We will implement a client-/server action pipelining system like it
existed in earlier prototypes, only now with expressions.

The JSON block format for actions was not the worst, but the composition
to pipelines wasn't optimal.

The expression language (now with map literals) is well suited to
express action pipelines in a nicely readable fashion.

Server-Side Actions
===================

We will have a pool of action implementations, which will transparently
cross the client-server gap and back.

we have to create the ability to execute actions in both contexts.

The server-side actions are defined as implementation of
de.quinscape.exceed.runtime.action.Action

/**
 * Implemented by the server-side of the unified action system.
 *
 * @param <M> action model type
 * @param <I> data input type.
 */
public interface Action<M extends ActionModel, I>
{
    Object execute(RuntimeContext runtimeContext, M model, I input);

    Class<M> getActionModelClass();

    Class<I> getInputClass();
}

The actions will automatically be picked up by component scan based on
implementing the interface.

Each action defines two classes that exist both as generics and as
runtime class values:

  * actionModelClass: Subclass of ActionModel that holds the
    configuration for the current action execution

  * inputClass: Runtime data to receive from the invoking side / a
    previous action.

The action implementation can then perform its job and can return a new
data object as next value in the action pipeline.

Action-Controller
=================

We create an action controller with two endpoints

  * POST: /action/{app}/{action}
  * GET: /action/{app}

The first one executes the action and the second one lists all server
actiosn in the application.

The actions are currently not really application dependent except for
the created runtime context.

In the future we will likely have application defined actions.

Action Expressions
==================

We use a specialized component expression to specifiy the actions in the
view components.

Expressions like:

	store().navigateTo({ view: 'home' })

are transformed into simplified js expressions

	[{'action':'store'},{'action':'navigateTo','view':'home'}]

Action Context
--------------

The expression can access the current component context, for example:

  * prop('name') : Provide the name prop of the current component
  * var('filter') : Provides the current value of the var 'filter' of
                    the current component
  * context: The current context if a parent component is a context
    providing component

The ViewComponent provides the context as variables "context", "props"
and "vars", which the transformer converts above syntax to.