Model: Add enum types
We already had a sketch for enum properties, but the idea of ad-hoc enum
definition by type param is not good, I think.

Better to have a formal definition of enum. An enum has an identfying
name (top-level model) and a values, which is a list of String values
valid for that enum, also defining ordinals for these values.

The idea is to always use the string values. For display purpose we will
us an i18n key "EnumType.CONSTANT". We might implement other high level
services based on enums.

For now, we define the EnumModel class and let the model composition
service compose these loaded from the extension resource
"models/domain/enum/".

We keep a map name -> enum in the application model.