Build/Tests: Babel 6, component testing
We switch to the new babel 6 and change all the transforms to the
equivalent list of plugins. We stop using gulp-espower and use
the babel transform instead.

Now we have power-assert and jsx working together in the tests.

Adds immutable.js support and some other helper libs.

Switch to react 0.14.6. Retire using react-with-addons. Add
"react-addons-*" dependencies if necessary.

Change shimed react package to just "react", change WebConfiguration.

Adds the react testing utils and jsdom for testing react components.

Update to current node and npm versions in pom.xml for jsdom.