ViewComponent: Improve compatibility, add test
We write a real test for our view component and the way it does view
composition and data injection.

We refactor some things for better testability.

We use a mix of React shallow rendering tests and full render tests with
jsdom. We can also use all methods of the react test helpers.

We create a module "test-dom" that sets up the standard jsdom testing
document containing just a div.root.

the module exports two methods:

    render(component, callback)

renders a component into the default jsdom
div.root and gives the callback to React as render callback for
validations.

    cleanup()

Unmounts the current component.

Component compatibility:
------------------------

We make sure to not modify the props from what is defined in the model
unless the component declaration opts into a certain behavior.

The component only receives a model atrbiute, if the component
descriptor has a "modelAware" : true property.

The component receives only a context if a contextKey or context
property are set.

If the "context" property is a string, the component will receive
the context as the prop corresping to the string value and not
as "context" property as it is by default.

This should make it much easier to integrate third party components
into our extended component model.

Module "security":
------------------

More testable by feeding the string separated role string from the
outside. Also make sure that the .init() can be called multiple times.