Build: Start conditional build
We introduce different compilation profiles based on prod and dev. We
use the nodejs NODE_ENV environment variable which we then line with
"babel-plugin-transform-node-env-inline" which improves on the old
module for that purpose in that it simplifies the whole evaluation
expression.

    if (process.env.NODE_ENV === "production")

gets transformed into

    if (true)

when transpiled with the NODE_ENV env variable set to "production".

These kind of simplified expressions than will be removed by uglify.

For conditional building however, this is not enough since the
optimization and potential removed of conditional require() calls
happens after browserify has determined all modules.

To remove the conditionally included modules and transitive
dependencies we need to also exclude the conditionally included modules
from the js build.

Outlook
-------

For now we just exclude the editor functionality for production builds
but in the future we should also exclude all the components in base
which are not needed in the current app.