HotReload: Handle aborts more gracefully
The exceptions that can occur while we are writing responses to the
client seem all to be client abort exceptions just caused by the browser
window being closed. I have never encountered another example.

To combat the flow of IOExceptions in the log that are not really
exceptions, we just quietly close the connection and log the exception
as DEBUG, just in case, to not swallow it completely