Refactor ViewComponents
The previous ViewComponent was a fixed react component that rendered the
current view from the current view model and view data at runtime.

The approach worked reasonably well and would have most likely been fast
enough for most cases, even though it redid a lot of work per request
and eval()ed attributes.

Creating the attribute evaluation context would have become more and
more complex and slower and slower and was also limited in a number of
ways.

I initially started out with dynamically generating the source code for
the views but stopped doing that because I thought it wouldn't work out
and could provide the contexts like we need them.

I've come to realize though that this estimate is wrong and combined
with another idea offers a far superior way of doing the views.

View Components: 3rd attempt
============================

We return to first rendering a component source code and then creating
the actual component from that by defining a Function which we can
give the necessary libs and context into.

We keep the way of handling contexts but with some additions.

In general we now again evaluate expressions by rendering them into the
view component source but now we use a server transform.

View Model Client JSON
----------------------

We begin to differentiate between the normal storage and parsing form of
views. When we cache the view JSON for client consumption we
additionally enhance the component tree with the precompiled
expressions, both for attributes and text child expressions.

The name,attrs,kids attributes of the component model stay exactly like
they are. For every attribute that is an expression, we transform that
expression from the normal syntax to what we need inside the view.

For example an expression like "props.foo" can turn into something like
"_v.root.kids[0].kids[2].kids[1].attrs.foo" etc.

For constant values, we also do inlining now so that that long format is
only used if the prop in question is a long string or an expression.

Component Descriptor changes
----------------------------

We change the component descriptor format a bit.

Context Provider:

If a component has a property "contextProvider" : true it will provide a
context value. In this case the child models of the model are not
rendered as normal kids but are wrapped inside a function prop

	function renderChildrenWithContext(context)
	{
		...
	}

which renders the children using the given context object.

The name of the context is "context" by default but can be defined by
using an attribute "var" : "name" on the component model of the context
provider.

For each context provider a renderChildrenWithContext closure is
rendered so that reusing the same context name is automatically scoped
right into the body of the context provider.

Context Consumer:

To receive a context object, the component descriptor must contain a
context propType declaration

{
    "components": {
        "Consumer": {
            "propTypes": {
                "myContext" : {
                    "context": true
                }
            }
        }
    }
}

Here we the the descriptor for a component "Consumer" which has
 context property named "myContext" which is marked as receiving a
context by having a "context" : true property.

The component will receive the context provided by first context
provider encountered going from the context consumer to its parents.

To receive another context we can give the name of the a context
as component model attribute with the same name ( "myContext" in this
case)

As you can see, a component can provide only one context but receive
contexts from multiple components.

Computed Consumer:

The concept of keyed consumer was now changed into computed consumers
which are more flexible.

{
       "KeyedConsumer": {
            "contextProvider" : true,
            "propTypes": {
                "name": { ... },
                "context" : {
                    "context": "context[props.name]"
                }
            }
        }
    }
}

Here we see an example for a computed consumer. For a computed consumer,
the context prop declaration needs to contain an expression instead of a
boolean.

The expression defines how to get from the parent context (always called
"context" in this kind of expression) to the context for the component.

In this case, we use the context property corresponding to the name prop
in the component model of the computed consumer.

Note that to provide the computed context to its own children, the
computed consumer also needs to be a context provider (following the
same scoping and "var" overriding rules).

Error-Handling
--------------

We set up react-transform-catch-errors to catch exceptions happening
during components during development without it interfering with the
rest of the view / the inpage editor.

We also wrap our generated ViewComponents with catchErrors by hand,
since we have no babel transformation at runtime.

We could also extend babel to work at runtime, but that would fatten up
our code bundles considerably and apart from this ErrorReporting case we
can handle everything fine just on the server side.

We introduce an exclusion of all files matching *.es5.js since we need
this exclusion at least for our ErrorReport lest
react-transform-catch-errors starts wrapping the ErrorReport in itself
while it is half-defined.

This also probvides an escape hatch should you ever find that one of the
babel transforms is useful in general but does something bad in a
specific case. Just name your module *.es5.js and write ES5 without JSX
and ES2015 features.

We activate react-transform-catch-errors when NODE_ENV is not set to
"production" and the env var NO_CATCH_ERRORS is not set.

view fixup

view fixup