Actions: Implement client side
Implement the client side to our action system and make some changes to
the server-side actions we had so far.

We introduce a general "systemInfo" data-block in the boostrapping JSP
view to which we add the names of all known server actions for now.
This replaces looking up the actions via AJAX at startup.

We move JSON conversion out of the default action registry and into the
application controller. We implement a default method to provide us
with all registered action models.

Action.js
=========

We implement an action service that registers bulk-required action
modules and can execute them.

It also offers a manual action registration method

actionService.register(actionName, handler, errorHandler)

to manually register client actions.

base is setup to automatically register

	src/main/js/action/**/*.js

as client-side actions.

For every module file an action is registered with the name being the
module name. The module is expected to export a handler function.

It will be invoked with the action model and the current data

	handler(actionModule, data)

The action can return either normal values or a promise. Non-promise
values are wrapped as promises and take part in the action service
promise chain.

The chain can of course either completely succeed or fail at one of the
promises. The default behaviour is fail the complete action
execution chain.

You can define a .catch property on the exported handler method to
handle failure of your action yourself. The error handler can also be
given as third parameter to actionService.register.

Action-Execution
----------------

actionService.execute(action, data)

returns a promise that resolves with the final data output after the
actions are executed.

action can either be a single action or an array of actions.

Action-Format
-------------

{
	action: "myAction",
	param: "param"
}

This corresponds to the ActionModel on the server-side (if it is a
server-side action) with the difference that the "action" property
declares the name of the action to parse the object into and gets
deleted before conversion.

Multiple actions can be called at once by using an array of actions.

The first action will be called with the given data, then the return
value of the first action will be given to the second etc.

The array can contain a mixture of client and server actions and will be
executed as one big promise chain.

We ass some tests for the action service.

We integrate a general throbber into the action system that will be
invoked after 3 seconds if the action(-chain) is not finished by then.