Bootstrap: Fix bootstrap resources
So far we've been service CSS with a dedicated controller and that is
not really working due to resources being referenced in the bootstrap
CSS.

We fix the issue by abandoning the style controller and making a special
case in the resource controller

	/res/style/{appName}.css

is used to serve the DSSed application styles for every application.
This works better because the new location lies within /res and has the
same relative paths as the CSS files in the /res/css/ folder.

We fix login and react-base references.

This restores styles to the login page.

Default App Problem
-------------------

From a setup point of view, the login page is a bit difficult in that it
does not belong to an application and has no application name associated
with it.

We fix that by taking the name of the current default app which works
for now but is not a long-term solution.

Maybe there should be a general special "admin" app that controls
deployment of all apps in that servlet context / database schema.